Begin to remove beaconing
