+++ app/outliers.py
@@ -210,10 +210,6 @@ def perform_analysis():
                 terms_analyzer = TermsAnalyzer(config_section_name=config_section_name)
                 analyzers.append(terms_analyzer)
 
-            elif config_section_name.startswith("beaconing_"):
-                beaconing_analyzer = BeaconingAnalyzer(config_section_name=config_section_name)
-                analyzers.append(beaconing_analyzer)
-
             elif config_section_name.startswith("word2vec_"):
                 word2vec_analyzer = Word2VecAnalyzer(config_section_name=config_section_name)
                 analyzers.append(word2vec_analyzer)
