+++ app/tests/unit_tests/files/beaconing_test_01.conf
@@ -37,11 +37,14 @@ beaconing_batch_eval_size=1000000
 ##############################
 # BEACONING - DUMMY TEST
 ##############################
-[beaconing_dummy_test]
+[terms_beaconing_dummy_test]
 es_query_filter=es_valid_query
 
 aggregator=BroFilter.server_name,BroFilter.id_orig_h,timestamp_day
 target=timestamp_hour
+target_count_method=within_aggregator
+trigger_on=low
+trigger_method=coeff_of_variation
 trigger_sensitivity=1
 min_target_buckets=2
