+++ app/tests/unit_tests/test_analyzer_terms.py
@@ -14,6 +14,8 @@ doc_without_outliers_test_whitelist_02_test_file = json.load(
     open("/app/tests/unit_tests/files/doc_without_outliers_test_whitelist_02.json"))
 doc_without_outliers_test_whitelist_03_test_file = json.load(
     open("/app/tests/unit_tests/files/doc_without_outliers_test_whitelist_03.json"))
+doc_without_outlier_test_file = json.load(open("/app/tests/unit_tests/files/doc_without_outlier.json"))
+doc_with_beaconing_outlier_test_file = json.load(open("/app/tests/unit_tests/files/doc_with_beaconing_outlier.json"))
 
 
 class TestTermsAnalyzer(unittest.TestCase):
@@ -47,3 +49,19 @@ class TestTermsAnalyzer(unittest.TestCase):
         analyzer.evaluate_model()
 
         self.assertEqual(len(analyzer.outliers), 2)
+
+    def test_evaluate_model_beaconing_simple_case(self):
+        self.test_settings.change_configuration_path("/app/tests/unit_tests/files/beaconing_test_01.conf")
+        analyzer = TermsAnalyzer("terms_beaconing_dummy_test")
+
+        doc_without_outlier = copy.deepcopy(doc_without_outlier_test_file)
+        expected_doc = copy.deepcopy(doc_with_beaconing_outlier_test_file)
+        # Add doc to the database
+        self.test_es.add_doc(doc_without_outlier)
+
+        # Make test (suppose that all doc match with the query)
+        analyzer.evaluate_model()
+
+        result = [elem for elem in es.scan()][0]
+        print(result["_source"])
+        self.assertEqual(result, expected_doc)
